requirements.txt
attrs
GitPython
multivolumefile
py7zr
tqdm
